move kafka configuration into separated package
