+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/Configuration.kt
@@ -1,6 +1,6 @@
 package github.com.hukuta94.delivery.api.startup.configuration
 
-import github.com.hukuta94.delivery.api.startup.configuration.adapter.KafkaConfiguration
+import github.com.hukuta94.delivery.api.startup.configuration.adapter.kafka.KafkaConfiguration
 import github.com.hukuta94.delivery.api.startup.configuration.adapter.http.ControllerConfiguration
 import github.com.hukuta94.delivery.api.startup.configuration.adapter.orm.OrmRepositoryConfiguration
 import github.com.hukuta94.delivery.api.startup.configuration.courier.CourierConfiguration
